Исправил id и всплывающее окно
