+++ src/components/modal-overlay/modal-overlay.jsx
@@ -1,22 +1,11 @@
-import ReactDOM from 'react-dom';
 import styles from './modal-overlay.module.css';
 import PropTypes from 'prop-types';
-
-const modalOverlay = document.getElementById("modal");
-
 function ModalOverlay({ onClick, children }) {
-    if (modalOverlay) {
-        return ReactDOM.createPortal(
-            (
-                <div className={styles.overlay} onClick={onClick}>
-                    {children}
-                </div>
-            ),
-            modalOverlay
-        );
-    } else {
-        return null;
-    }
+    return (
+        <div className={styles.overlay} onClick={onClick}>
+            {children}
+        </div>
+    );
 }
 
 ModalOverlay.propTypes = {
