+++ src/components/modal/modal.jsx
@@ -3,8 +3,10 @@ import ModalOverlay from '../modal-overlay/modal-overlay';
 import styles from './modal.module.css';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 import PropTypes from 'prop-types';
+import ReactDOM from 'react-dom';
 
 function Modal({ onClick, children }) {
+    const modalOverlay = document.getElementById("modal");
 
     useEffect(() => {
         const handleKeyPress = (event) => {
@@ -19,19 +21,23 @@ function Modal({ onClick, children }) {
             document.removeEventListener('keydown', handleKeyPress);
         };
     }, [onClick]);
-
-    return (
-        <ModalOverlay onClick={onClick}>
-            <div className={`${styles.modal} pt-10`}>
-                <div className={styles.header}>
-                    <button className={`${styles.closeIcon} mt-15 mr-10`} onClick={onClick}>
-                        <CloseIcon type="primary" />
-                    </button>
+    if (modalOverlay) {
+        return ReactDOM.createPortal(
+            <ModalOverlay onClick={onClick}>
+                <div className={`${styles.modal} pt-10`}>
+                    <div className={styles.header}>
+                        <button className={`${styles.closeIcon} mt-15 mr-10`} onClick={onClick}>
+                            <CloseIcon type="primary" />
+                        </button>
+                    </div>
+                    {children}
                 </div>
-                {children}
-            </div>
-        </ModalOverlay>
-    );
+            </ModalOverlay>,
+            modalOverlay
+        );
+    } else {
+        return null;
+    }
 }
 
 ModalOverlay.propTypes = {
